pyproject.toml
[tool.poetry]
name = "helsinki-callscript"
version = "1.2.2"
description = "A python script to trigger calls on a desk phone connected to 3cx"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
websocket-client = "^1.8.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
skip-string-normalization = false
